pyproject.toml
[project]
name = "bong_kindle"
version = "0.0.1"
description = "A node based flow editor for AI workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "fire>=0.7.0",
    "pypub3>=2.0.9",
    "requests>=2.32.3",
    "rich>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "uv>=0.4.20",
    "pytest>=8.3.3",
    "isort>=5.13.2",
    "black>=24.10.0",
    "ruff>=0.6.9",
]

[project.scripts]
bong_kindle = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bong_kindle"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["I"]

[tool.ruff.lint.isort]
known-first-party = ["bong_kindle"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
combine-as-imports = true
